refact: rendering

+++ src/ui/mod.rs
@@ -38,7 +38,7 @@ pub trait Component: Send {
         }
     }
 
-    fn cursor_position(&self) -> Option<(u16, u16)> {
+    fn cursor_position(&self) -> Option<(usize, usize)> {
         None
     }
 }
